:zap: 기안번호 자동 생성 연동하여 미리 생성하지 않아도 휴가 신청 가능

+++ src/main/java/com/semi/lynk/function/human/model/calendar/VacationApplicationDTO.java
@@ -36,7 +36,7 @@ public class VacationApplicationDTO {
     private int scheduleType;           // 스케쥴 타입
     private String scheduleNote;        // 스케쥴 메모 (사 유)
 
-    private int draftNo;
+    private Long draftNo;
     private int employeeNo;
 
 }
